requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
greenlet==3.1.0

# Auth & Security
pyjwt==2.9.0
bcrypt==4.2.0
cryptography==43.0.0

# HTTP client (for Railway API)
httpx==0.27.0

# Stripe
stripe==10.0.0

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.9.0
pydantic-settings==2.5.0
